Simplification
I've removed all instances of the collision detection from this code to
make it easier to understand.

This code simply reproduces the point cloud across the entire array of
kinects. In the future it will incorporate a moving portion that allows
you to find out where the different kinects are.